pyproject.toml
[project]
name = "paper-online"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi[standard]>=0.128.0",
    "openai>=2.16.0",
    "psycopg[binary]>=3.2.10",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "sse-starlette>=3.2.0",
    "tiktoken>=0.12.0",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
